+++ src/main/java/koreatech/in/dto/upload/response/UploadFilesResponse.java
@@ -1,4 +1,4 @@
-package koreatech.in.dto.upload;
+package koreatech.in.dto.upload.response;
 
 import io.swagger.annotations.ApiModelProperty;
 import java.util.List;
